Made Changes Based on the Review

+++ extras/perf_test_cgal.cpp
@@ -40,7 +40,7 @@ typedef CGAL::SM_Vertex_index Vertex;
 void manifoldToCGALSurfaceMesh(Manifold &manifold, TriangleMesh &cgalMesh) {
   auto maniMesh = manifold.GetMesh();
 
-  const int n = maniMesh.vertPos.size();
+  const size_t n = maniMesh.vertPos.size();
   std::vector<Vertex> vertices(n);
   for (size_t i = 0; i < n; i++) {
     auto &vert = maniMesh.vertPos[i];
